Fix issues with multiblock structure preview
