+++ src/main/java/net/oktawia/crazyae2addons/entities/EntropyCradleControllerBE.java
@@ -15,7 +15,6 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.mojang.logging.LogUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
@@ -33,8 +32,6 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.energy.EnergyStorage;
 import net.minecraftforge.energy.IEnergyStorage;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -43,22 +40,31 @@ import net.oktawia.crazyae2addons.defs.regs.CrazyBlockEntityRegistrar;
 import net.oktawia.crazyae2addons.defs.regs.CrazyBlockRegistrar;
 import net.oktawia.crazyae2addons.defs.regs.CrazyMenuRegistrar;
 import net.oktawia.crazyae2addons.menus.EntropyCradleControllerMenu;
-import net.oktawia.crazyae2addons.menus.PenroseControllerMenu;
-import net.oktawia.crazyae2addons.misc.CachedBlockInfo;
+import net.oktawia.crazyae2addons.renderer.preview.PreviewInfo;
 import net.oktawia.crazyae2addons.misc.CradleRecipes;
-import net.oktawia.crazyae2addons.misc.EntropyCradlePreviewRenderer;
 import net.oktawia.crazyae2addons.misc.EntropyCradleValidator;
+import net.oktawia.crazyae2addons.renderer.preview.Previewable;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
-public class EntropyCradleControllerBE extends AENetworkInvBlockEntity implements IGridTickable, MenuProvider {
+public class EntropyCradleControllerBE extends AENetworkInvBlockEntity implements Previewable, IGridTickable, MenuProvider {
 
     public EntropyCradleValidator validator;
     public int MAX_ENERGY = 600_000_000;
     public IEnergyStorage storedEnergy;
 
-    public List<CachedBlockInfo> ghostCache = null;
+    private PreviewInfo previewInfo = null;
+
+    @Override
+    public PreviewInfo getPreviewInfo() {
+        return previewInfo;
+    }
+
+    @Override
+    public void setPreviewInfo(PreviewInfo info) {
+        this.previewInfo = info;
+    }
 
     public boolean preview = false;
