+++ src/main/java/net/oktawia/crazyae2addons/entities/PatternManagementUnitControllerBE.java
@@ -27,16 +27,15 @@ import net.oktawia.crazyae2addons.defs.regs.CrazyBlockEntityRegistrar;
 import net.oktawia.crazyae2addons.defs.regs.CrazyBlockRegistrar;
 import net.oktawia.crazyae2addons.defs.regs.CrazyMenuRegistrar;
 import net.oktawia.crazyae2addons.menus.PatternManagementUnitControllerMenu;
-import net.oktawia.crazyae2addons.misc.CachedBlockInfo;
-import net.oktawia.crazyae2addons.misc.PatternManagementUnitPreviewRenderer;
+import net.oktawia.crazyae2addons.renderer.preview.PreviewInfo;
 import net.oktawia.crazyae2addons.misc.PatternManagementUnitValidator;
+import net.oktawia.crazyae2addons.renderer.preview.Previewable;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
-public class PatternManagementUnitControllerBE extends AENetworkBlockEntity implements IGridTickable, InternalInventoryHost, MenuProvider {
+public class PatternManagementUnitControllerBE extends AENetworkBlockEntity implements Previewable, IGridTickable, InternalInventoryHost, MenuProvider {
 
     public PatternManagementUnitControllerBE controller;
     public PatternManagementUnitValidator validator;
@@ -50,7 +49,17 @@ public class PatternManagementUnitControllerBE extends AENetworkBlockEntity impl
     public boolean init;
     public boolean active;
 
-    public List<CachedBlockInfo> ghostCache = null;
+    private PreviewInfo previewInfo = null;
+
+    @Override
+    public PreviewInfo getPreviewInfo() {
+        return previewInfo;
+    }
+
+    @Override
+    public void setPreviewInfo(PreviewInfo info) {
+        this.previewInfo = info;
+    }
 
     public static final Set<PatternManagementUnitControllerBE> CLIENT_INSTANCES = new HashSet<>();
