+++ src/main/java/net/oktawia/crazyae2addons/entities/PenroseControllerBE.java
@@ -40,6 +40,8 @@ import net.oktawia.crazyae2addons.defs.regs.CrazyItemRegistrar;
 import net.oktawia.crazyae2addons.defs.regs.CrazyMenuRegistrar;
 import net.oktawia.crazyae2addons.menus.PenroseControllerMenu;
 import net.oktawia.crazyae2addons.misc.*;
+import net.oktawia.crazyae2addons.renderer.preview.PreviewInfo;
+import net.oktawia.crazyae2addons.renderer.preview.Previewable;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -48,7 +50,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
-public class PenroseControllerBE extends AENetworkInvBlockEntity implements MenuProvider, IUpgradeableObject, IGridTickable {
+public class PenroseControllerBE extends AENetworkInvBlockEntity implements Previewable, MenuProvider, IUpgradeableObject, IGridTickable {
     public static final Set<PenroseControllerBE> CLIENT_INSTANCES = new HashSet<>();
     public boolean energyMode = false;
     public boolean preview = false;
@@ -56,9 +58,19 @@ public class PenroseControllerBE extends AENetworkInvBlockEntity implements Menu
     private int myticks = 0;
     public EnergyStorage energyStorage = new EnergyStorage(Integer.MAX_VALUE, 0, Integer.MAX_VALUE, 0);
 
-    public List<CachedBlockInfo> ghostCache = null;
-
     public int cachedTier = -1;
+
+    private PreviewInfo previewInfo = null;
+
+    @Override
+    public PreviewInfo getPreviewInfo() {
+        return previewInfo;
+    }
+
+    @Override
+    public void setPreviewInfo(PreviewInfo info) {
+        this.previewInfo = info;
+    }
 
     private final LazyOptional<IEnergyStorage> energyCap = LazyOptional.of(() -> new IEnergyStorage() {
         @Override public int getEnergyStored() {
