+++ src/main/java/net/oktawia/crazyae2addons/renderer/preview/EnergyStoragePreviewRenderer.java
@@ -1,42 +1,38 @@
-package net.oktawia.crazyae2addons.misc;
+package net.oktawia.crazyae2addons.renderer.preview;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
-import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.RenderLevelStageEvent;
 import net.minecraftforge.client.event.RenderLevelStageEvent.Stage;
-import net.minecraftforge.client.model.data.ModelData;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.oktawia.crazyae2addons.CrazyAddons;
-import net.oktawia.crazyae2addons.blocks.EntropyCradle;
-import net.oktawia.crazyae2addons.blocks.EntropyCradleCapacitor;
-import net.oktawia.crazyae2addons.entities.EntropyCradleControllerBE;
+import net.oktawia.crazyae2addons.blocks.EnergyStorageFrame;
+import net.oktawia.crazyae2addons.entities.EnergyStorageControllerBE;
+import net.oktawia.crazyae2addons.renderer.PreviewRenderer;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 @Mod.EventBusSubscriber(modid = CrazyAddons.MODID, value = Dist.CLIENT)
-public class EntropyCradlePreviewRenderer {
+public class EnergyStoragePreviewRenderer {
 
     @SubscribeEvent
     public static void onRender(RenderLevelStageEvent event) {
-        if (event.getStage() != Stage.AFTER_SOLID_BLOCKS) return;
+        if (event.getStage() != Stage.AFTER_TRANSLUCENT_BLOCKS) return;
 
         Minecraft mc = Minecraft.getInstance();
         if (mc.level == null || mc.player == null) return;
 
-        for (EntropyCradleControllerBE controller : EntropyCradleControllerBE.CLIENT_INSTANCES) {
+        for (EnergyStorageControllerBE controller : EnergyStorageControllerBE.CLIENT_INSTANCES) {
             if (!controller.preview) continue;
 
             BlockPos origin = controller.getBlockPos();
@@ -46,15 +42,15 @@ public class EntropyCradlePreviewRenderer {
                     .getValue(BlockStateProperties.HORIZONTAL_FACING)
                     .getOpposite();
 
-            if (controller.ghostCache == null) {
+            if (controller.getPreviewInfo() == null) {
                 rebuildCache(controller, facing);
             }
 
-            PreviewRenderer.render(controller.ghostCache, event);
+            PreviewRenderer.render(controller.getPreviewInfo(), event);
         }
     }
 
-    private static void rebuildCache(EntropyCradleControllerBE controller, Direction facing) {
+    private static void rebuildCache(EnergyStorageControllerBE controller, Direction facing) {
         Minecraft mc = Minecraft.getInstance();
         BlockRenderDispatcher blockRenderer = mc.getBlockRenderer();
 
@@ -66,12 +62,11 @@ public class EntropyCradlePreviewRenderer {
         int originZ = validator.getOriginZ();
 
         BlockPos origin = controller.getBlockPos();
-        List<CachedBlockInfo> cache = new ArrayList<>();
-
         int height = layers.size();
         int sizeZ = layers.get(0).size();
         int sizeX = layers.get(0).get(0).split(" ").length;
 
+        var blockInfos = new ArrayList<PreviewInfo.BlockInfo>();
         for (int y = 0; y < height; y++) {
             List<String> layer = layers.get(y);
             for (int z = 0; z < sizeZ; z++) {
@@ -84,10 +79,8 @@ public class EntropyCradlePreviewRenderer {
                     if (blocks.isEmpty()) continue;
 
                     BlockState state = blocks.get(0).defaultBlockState();
-                    if (state.getBlock() instanceof EntropyCradle){
-                        state = state.setValue(EntropyCradle.FORMED, true);
-                    } else if (state.getBlock() instanceof EntropyCradleCapacitor){
-                        state = state.setValue(EntropyCradleCapacitor.FORMED, true);
+                    if (state.getBlock() instanceof EnergyStorageFrame){
+                        state = state.setValue(EnergyStorageFrame.FORMED, true);
                     }
                     int relX = x - originX;
                     int relY = y - originY;
@@ -96,12 +89,12 @@ public class EntropyCradlePreviewRenderer {
                     BlockPos pos = origin.offset(offset.getX(), relY, offset.getZ());
 
                     BakedModel model = blockRenderer.getBlockModel(state);
-                    cache.add(new CachedBlockInfo(pos, state, model));
+                    blockInfos.add(new PreviewInfo.BlockInfo(pos, state, model));
                 }
             }
         }
 
-        controller.ghostCache = cache;
+        controller.setPreviewInfo(new PreviewInfo(blockInfos));
     }
 
     private static BlockPos rotateOffset(int x, int z, Direction facing) {
