+++ src/main/java/net/oktawia/crazyae2addons/renderer/preview/EntropyCradlePreviewRenderer.java
@@ -1,4 +1,4 @@
-package net.oktawia.crazyae2addons.misc;
+package net.oktawia.crazyae2addons.renderer.preview;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
@@ -14,23 +14,24 @@ import net.minecraftforge.client.event.RenderLevelStageEvent.Stage;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.oktawia.crazyae2addons.CrazyAddons;
-import net.oktawia.crazyae2addons.entities.PatternManagementUnitControllerBE;
+import net.oktawia.crazyae2addons.blocks.EntropyCradle;
+import net.oktawia.crazyae2addons.blocks.EntropyCradleCapacitor;
+import net.oktawia.crazyae2addons.entities.EntropyCradleControllerBE;
+import net.oktawia.crazyae2addons.renderer.PreviewRenderer;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @Mod.EventBusSubscriber(modid = CrazyAddons.MODID, value = Dist.CLIENT)
-public class PatternManagementUnitPreviewRenderer {
+public class EntropyCradlePreviewRenderer {
 
     @SubscribeEvent
     public static void onRender(RenderLevelStageEvent event) {
-        if (event.getStage() != Stage.AFTER_SOLID_BLOCKS) return;
+        if (event.getStage() != Stage.AFTER_TRANSLUCENT_BLOCKS) return;
 
         Minecraft mc = Minecraft.getInstance();
         if (mc.level == null || mc.player == null) return;
 
-        for (PatternManagementUnitControllerBE controller : PatternManagementUnitControllerBE.CLIENT_INSTANCES) {
+        for (EntropyCradleControllerBE controller : EntropyCradleControllerBE.CLIENT_INSTANCES) {
             if (!controller.preview) continue;
 
             BlockPos origin = controller.getBlockPos();
@@ -40,15 +41,15 @@ public class PatternManagementUnitPreviewRenderer {
                     .getValue(BlockStateProperties.HORIZONTAL_FACING)
                     .getOpposite();
 
-            if (controller.ghostCache == null) {
+            if (controller.getPreviewInfo() == null) {
                 rebuildCache(controller, facing);
             }
 
-            PreviewRenderer.render(controller.ghostCache, event);
+            PreviewRenderer.render(controller.getPreviewInfo(), event);
         }
     }
 
-    private static void rebuildCache(PatternManagementUnitControllerBE controller, Direction facing) {
+    private static void rebuildCache(EntropyCradleControllerBE controller, Direction facing) {
         Minecraft mc = Minecraft.getInstance();
         BlockRenderDispatcher blockRenderer = mc.getBlockRenderer();
 
@@ -60,12 +61,11 @@ public class PatternManagementUnitPreviewRenderer {
         int originZ = validator.getOriginZ();
 
         BlockPos origin = controller.getBlockPos();
-        List<CachedBlockInfo> cache = new ArrayList<>();
-
         int height = layers.size();
         int sizeZ = layers.get(0).size();
         int sizeX = layers.get(0).get(0).split(" ").length;
 
+        var blockInfos = new ArrayList<PreviewInfo.BlockInfo>();
         for (int y = 0; y < height; y++) {
             List<String> layer = layers.get(y);
             for (int z = 0; z < sizeZ; z++) {
@@ -78,6 +78,11 @@ public class PatternManagementUnitPreviewRenderer {
                     if (blocks.isEmpty()) continue;
 
                     BlockState state = blocks.get(0).defaultBlockState();
+                    if (state.getBlock() instanceof EntropyCradle){
+                        state = state.setValue(EntropyCradle.FORMED, true);
+                    } else if (state.getBlock() instanceof EntropyCradleCapacitor){
+                        state = state.setValue(EntropyCradleCapacitor.FORMED, true);
+                    }
                     int relX = x - originX;
                     int relY = y - originY;
                     int relZ = z - originZ;
@@ -85,21 +90,21 @@ public class PatternManagementUnitPreviewRenderer {
                     BlockPos pos = origin.offset(offset.getX(), relY, offset.getZ());
 
                     BakedModel model = blockRenderer.getBlockModel(state);
-                    cache.add(new CachedBlockInfo(pos, state, model));
+                    blockInfos.add(new PreviewInfo.BlockInfo(pos, state, model));
                 }
             }
         }
 
-        controller.ghostCache = cache;
+        controller.setPreviewInfo(new PreviewInfo(blockInfos));
     }
 
     private static BlockPos rotateOffset(int x, int z, Direction facing) {
         return switch (facing) {
             case NORTH -> new BlockPos(x, 0, z);
             case SOUTH -> new BlockPos(-x, 0, -z);
-            case WEST  -> new BlockPos(z, 0, -x);
-            case EAST  -> new BlockPos(-z, 0, x);
-            default    -> BlockPos.ZERO;
+            case WEST -> new BlockPos(z, 0, -x);
+            case EAST -> new BlockPos(-z, 0, x);
+            default -> BlockPos.ZERO;
         };
     }
 }
