+++ src/main/java/net/oktawia/crazyae2addons/renderer/preview/Previewable.java
@@ -1,0 +1,7 @@
+package net.oktawia.crazyae2addons.renderer.preview;
+
+public interface Previewable {
+    PreviewInfo getPreviewInfo();
+
+    void setPreviewInfo(PreviewInfo previewInfo);
+}
